extract generic types
